WIP: Reorganize types
